DRY fix for deferred route decorator.
